Losen type definitions Int64 -> Int
Losen type definitions to make code work on 32-bit devices also. In
general, in function definitions we should use either Int which is
typealias to Int32 or Int64 depending machine architecture, or abstract
type Integer.

+++ src/elements_lagrange.jl
@@ -19,7 +19,7 @@ function get_integration_order(::Poi1)
     return 1
 end
 
-function get_integration_points(::Poi1, ::Int64)
+function get_integration_points(::Poi1, ::Int)
     return [ (1.0, (0.0, )) ]
 end
